feat: add toggle member functionality

+++ database/migrations/2023_01_07_103912_create_members_table.php
@@ -55,7 +55,7 @@ return new class extends Migration
             $table->boolean('delivery_status')->default(Member::DELIVERY_STATUS_DEFAULT);
 
             $table->string('email')->unique();
-            $table->integer('mobile')->unique();
+            $table->bigInteger('mobile')->unique();
             $table->integer('mobile_code')->nullable();
             $table->string('password');
